+++ R/validation.R
@@ -41,7 +41,7 @@ validate <- function(object, participantColumn = FALSE, validateRegression = FAL
     start.time <- Sys.time()
     
     # Make resampled model
-    temp_object[[ii]] <- prepateValidationRun(object)
+    temp_object[[ii]] <- prepareValidationRun(object)
       
     
     # Rotate new loadings/scores to the original model
@@ -53,7 +53,14 @@ validate <- function(object, participantColumn = FALSE, validateRegression = FAL
     cat("--- Used ",round(time_mean[ii],2)," seconds. Est. time remaining: ",round((object$nValRuns-ii)*mean(time_mean),2)," seconds \n")
   }
   
+  if(grepl("permutation",object$validationMethod)){
+    cat("- Calculates P values...\n")
+    object <- getPermutationPValues(object, objectlist = temp_object)
+  }
+  
+  cat("- Calculates percentiles for score and loading...\n")
   object <- getValidationPercentiles(object, objectlist = temp_object)
+  
   if(object$keepValidationObjects){
     object$validation$temp_objects <- temp_object
   }
@@ -349,7 +356,7 @@ getRegressionPredictions <- function(object){
 #'
 #' @param object An ALASCA object
 #' @return An ALASCA object
-prepateValidationRun <- function(object){
+prepareValidationRun <- function(object){
   if(object$validationMethod == "loo"){
     # Use leave-one-out validation
     selectedParts <- data.frame()
@@ -380,17 +387,18 @@ prepateValidationRun <- function(object){
     }
   }else if(object$validationMethod == "permutation2"){
     # Validation by permutation
+    # Randomize samples
     parts <- data.frame(
       time = object$df$time,
       group = object$df$group,
-      ID = object$df[,colnames(object$df) == object$participantColumn]
+      ID = object$df[,ID]
     )
     parts <- parts[!duplicated(parts),]
     times <- parts$time
     groups <- parts$group
     
     # sample id
-    parts_orig <- paste0(object$df[,partColumn], object$df$time, object$df$group)
+    parts_orig <- paste0(object$df[,ID], object$df$time, object$df$group)
     u_parts_orig <- unique(parts_orig)
     
     temp_object <- object
@@ -403,34 +411,122 @@ prepateValidationRun <- function(object){
       groups <- groups[-rIDGroup]
     }
     temp_object <- ALASCA(validationObject = temp_object,
-                          validationParticipants = !is.na(object$df[,partColumn]))
+                          validationParticipants = !is.na(object$df[,ID]))
   }else if(object$validationMethod == "permutation"){
     # Validation by permutation
+    # Randomize individuals across groups
+    # Randomize time within individual
     parts_g <- data.frame(
       group = object$df$group,
-      ID = object$df[,colnames(object$df) == object$participantColumn]
+      ID = object$df[,ID]
     )
     parts_g <- parts_g[!duplicated(parts_g),]
     groups <- parts_g$group
     
     # sample id
-    parts_orig <- unique(paste0(object$df[,partColumn]))
+    parts_orig <- unique(paste0(object$df[,ID]))
     
     temp_object <- object
-    for(i in unique(object$df[,partColumn])){
+    for(i in unique(object$df[,ID])){
       rIDGroup <- sample(seq_along(groups), 1)
-      temp_object$dfRaw$group[temp_object$df[,partColumn] == i] <- groups[rIDGroup]
+      temp_object$dfRaw$group[temp_object$df[,ID] == i] <- groups[rIDGroup]
       groups <- groups[-rIDGroup]
-      times <- unique(temp_object$dfRaw$time[temp_object$df[,partColumn] == i])
-      for(j in unique(temp_object$dfRaw$time[temp_object$df[,partColumn] == i])){
+      times <- unique(temp_object$dfRaw$time[temp_object$df[,ID] == i])
+      for(j in unique(temp_object$dfRaw$time[temp_object$df[,ID] == i])){
         rIDTime <- sample(seq_along(times), 1)
-        temp_object$dfRaw$time[temp_object$df[,partColumn] == i & temp_object$df$time == j] <- times[rIDTime]
+        temp_object$dfRaw$time[temp_object$df[,ID] == i & temp_object$df$time == j] <- times[rIDTime]
         times <- times[-rIDTime]
       }
     }
     temp_object <- ALASCA(validationObject = temp_object,
-                          validationParticipants = !is.na(object$df[,partColumn]))
+                          validationParticipants = !is.na(object$df[,ID]))
   }
   
   return(temp_object)
+}
+
+#' Get P values
+#'
+#' This function calculates P values
+#'
+#' @param target The full-model ALASCA object
+#' @param objectlist List of permutated ALASCA objects
+#' @return An ALASCA object
+getPermutationPValues <- function(target, objectlist){
+  
+  if(target$separateTimeAndGroup){
+    # Calculate the Frobenius sum squares of the first l principal components
+    SS_f_time <- Reduce(rbind,lapply(seq_along(objectlist), function(i){
+      tmp <- reshape2::melt(objectlist[[i]]$pca$score$time, id.vars = c("time"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(objectlist[[i]]$pca$loading$explained$time),], value~time, FUN = function(x) sum(x^2))
+      tmp2$model <- i
+      tmp2
+    }))
+    SS_time <- Reduce(rbind,lapply(1, function(i){
+      tmp <- reshape2::melt(target$pca$score$time, id.vars = c("time"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(target$pca$loading$explained$time),], value~time, FUN = function(x) sum(x^2))
+      tmp2
+    }))
+    SS_f_group <- Reduce(rbind,lapply(seq_along(objectlist), function(i){
+      tmp <- reshape2::melt(objectlist[[i]]$pca$score$group, id.vars = c("time","group"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(objectlist[[i]]$pca$loading$explained$group),], value~group+time, FUN = function(x) sum(x^2))
+      tmp2$model <- i
+      tmp2
+    }))
+    SS_group <- Reduce(rbind,lapply(1, function(i){
+      tmp <- reshape2::melt(target$pca$score$group, id.vars = c("time","group"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(target$pca$loading$explained$group),], value~group+time, FUN = function(x) sum(x^2))
+      tmp2
+    }))
+    
+    pvals_time <- Reduce(rbind,lapply(unique(SS_time$time), function(x){
+      data.frame(
+        p.value = sum( SS_f_time$value[SS_f_time$time == x] >= SS_time$value[SS_time$time == x] ) / sum(SS_f_time$time == x),
+        effect = x,
+        nRuns = target$nValRuns
+      )
+    }))
+    
+    pvals_group <- Reduce(rbind,lapply(unique(paste(SS_group$time,SS_group$group)), function(x){
+      data.frame(
+        p.value = sum( SS_f_group$value[paste(SS_f_group$time,SS_f_group$group) == x] >= SS_group$value[paste(SS_group$time,SS_group$group) == x] ) / sum(paste(SS_f_group$time,SS_f_group$group) == x),
+        effect = x,
+        nRuns = target$nValRuns
+      )
+    }))
+    
+    pvals <- rbind(pvals_time,pvals_group)
+    
+  }else{
+    # Calculate the Frobenius sum squares of the first l principal components
+    SS_f <- Reduce(rbind,lapply(seq_along(objectlist), function(i){
+      tmp <- reshape2::melt(objectlist[[i]]$pca$score$time, id.vars = c("time","group"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(objectlist[[i]]$pca$loading$explained$time),], value~group+time, FUN = function(x) sum(x^2))
+      tmp2$model <- i
+      tmp2
+    }))
+    SS <- Reduce(rbind,lapply(1, function(i){
+      tmp <- reshape2::melt(target$pca$score$time, id.vars = c("time","group"))
+      tmp$variable <- as.numeric(gsub("PC","",tmp$variable))
+      tmp2 <- aggregate(data=tmp[tmp$variable %in% getRelevantPCs(target$pca$loading$explained$time),], value~group+time, FUN = function(x) sum(x^2))
+      tmp2
+    }))
+    pvals <- Reduce(rbind,lapply(unique(paste(SS$time,SS$group)), function(x){
+      data.frame(
+        p.value = sum( SS_f$value[paste(SS_f$time,SS_f$group) == x] >= SS$value[paste(SS$time,SS$group) == x] ) / sum(paste(SS_f$time,SS_f$group) == x),
+        effect = x,
+        nRuns = target$nValRuns
+      )
+    }))
+  }
+  
+  min_p_value <- 1/pvals$nRuns[1]
+  pvals$p.value[pvals$p.value < min_p_value] <- min_p_value
+  target$pvals <- pvals
+  return(target)
 }